BackEnd/requirements.txt
alembic==1.8.1
anyio==3.6.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
peewee==3.15.1
pyasn1==0.4.8
pydantic==1.9.1
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3